Official docs link of state to latest version
The existing link of official docs for state hook is from the previous website of React, the one i proposed is from the latest documentation of React although it is in beta, the documentation required for state hook has been completed. I hope I am committing against the newer version of the roadmap. Thank you !